requirements.txt
moviepy
Pillow
torch
openai-whisper
